pyproject.toml
[tool.poetry]
name = "afisha"
version = "0.1.0"
description = "Event management API"
authors = ["Whoagir <whoagir@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
django = ">=5.2.1,<6.0.0"
gunicorn = "^21.2.0"
celery = "5.3.6"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.1.0"
dj-database-url = "^2.3.0"
djangorestframework = "^3.15.0"
django-filter = "^23.5"
flower = "^1.2.0"
setuptools = "^68.2"
redis = "^5.0.1"
djangorestframework-simplejwt = "^5.5.0"
drf-spectacular = "^0.27.0"
django-cors-headers = "^4.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
